refactor(registration-react): Rename function-component of FormDistributionApi to FormDistributionApiPure and final wrapped component to FormDistributionApi
