+++ applications/registration-react/src/pages/dataset-registration-page/form-distribution-api/configured-form-distribution-api.jsx
@@ -1,7 +1,7 @@
 import { reduxForm } from 'redux-form';
 import _throttle from 'lodash/throttle';
 
-import { FormDistributionAPI } from './form-distribution-api';
+import { FormDistributionApiPure } from './form-distribution-api-pure';
 import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
@@ -10,5 +10,5 @@ const config = {
 };
 
 export const ConfiguredFormDistributionAPI = reduxForm(config)(
-  FormDistributionAPI
+  FormDistributionApiPure
 );
